Service repository

+++ prestador_de_servico/lib/app/repositories/service/sqflite_service_repository.dart
@@ -1,7 +1,11 @@
+import 'dart:io';
 import 'package:prestador_de_servico/app/repositories/config/sqflite_config.dart';
 import 'package:prestador_de_servico/app/models/service/service.dart';
 import 'package:prestador_de_servico/app/models/service/service_adapter.dart';
 import 'package:prestador_de_servico/app/repositories/service/service_repository.dart';
+import 'package:prestador_de_servico/app/shared/either/either.dart';
+import 'package:prestador_de_servico/app/shared/either/either_extension.dart';
+import 'package:prestador_de_servico/app/shared/failure/failure.dart';
 import 'package:replace_diacritic/replace_diacritic.dart';
 import 'package:sqflite/sqflite.dart';
 
@@ -11,17 +15,27 @@ class SqfliteServiceRepository implements ServiceRepository {
 
   SqfliteServiceRepository({this.database});
 
-  Future<void> _initDatabase() async {
+  Future<Either<Failure, Unit>> _initDatabase() async {
     SqfliteConfig sqfliteConfig = SqfliteConfig();
-    database ??= await sqfliteConfig.getDatabase();
-    if (servicesTable.isEmpty) {
-      servicesTable = sqfliteConfig.services;
+    try {
+      database ??= await sqfliteConfig.getDatabase();
+      if (servicesTable.isEmpty) {
+        servicesTable = sqfliteConfig.services;
+      }
+      return Either.right(unit);
+    } on DatabaseException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao acessar banco de dados local: $e'));
+    } on FileSystemException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao acessar arquivo de banco de dados local: ${e.message}'));
     }
   }
 
   @override
-  Future<List<Service>> getAll() async {
-    await _initDatabase();
+  Future<Either<Failure, List<Service>>> getAll() async {
+    final getDbEither = await _initDatabase();
+    if (getDbEither.isLeft) {
+      return Either.left(getDbEither.left);
+    }
 
     String selectCommand = ""
         "SELECT "
@@ -36,19 +50,23 @@ class SqfliteServiceRepository implements ServiceRepository {
         "FROM "
         "$servicesTable ser";
 
-    List<Map> servicesMap = await database!.rawQuery(selectCommand);
-
-    List<Service> services = servicesMap
-        .map((service) => ServiceAdapter.fromSqflite(map: service))
-        .toList();
-
-    return services;
+    try {
+      final servicesMap = await database!.rawQuery(selectCommand);
+      final services = servicesMap.map((service) => ServiceAdapter.fromSqflite(map: service)).toList();
+      return Either.right(services);
+    } on DatabaseException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao capturar dados locais: $e'));
+    } on FileSystemException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao capturar dados locais: ${e.message})'));
+    }
   }
 
   @override
-  Future<List<Service>> getByServiceCategoryId(
-      {required String serviceCategoryId}) async {
-    await _initDatabase();
+  Future<Either<Failure, List<Service>>> getByServiceCategoryId({required String serviceCategoryId}) async {
+    final getDbEither = await _initDatabase();
+    if (getDbEither.isLeft) {
+      return Either.left(getDbEither.left);
+    }
 
     String selectCommand = ""
         "SELECT "
@@ -64,19 +82,24 @@ class SqfliteServiceRepository implements ServiceRepository {
         "$servicesTable ser "
         "WHERE ser.serviceCategoryId = ?";
 
-    List params = [serviceCategoryId];
-    List<Map> servicesMap = await database!.rawQuery(selectCommand, params);
-
-    List<Service> services = servicesMap
-        .map((service) => ServiceAdapter.fromSqflite(map: service))
-        .toList();
-
-    return services;
+    try {
+      final params = [serviceCategoryId];
+      final servicesMap = await database!.rawQuery(selectCommand, params);
+      final services = servicesMap.map((service) => ServiceAdapter.fromSqflite(map: service)).toList();
+      return Either.right(services);
+    } on DatabaseException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao capturar dados locais: $e'));
+    } on FileSystemException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao capturar dados locais: ${e.message})'));
+    }
   }
 
   @override
-  Future<Service> getById({required String id}) async {
-    await _initDatabase();
+  Future<Either<Failure, Service>> getById({required String id}) async {
+    final getDbEither = await _initDatabase();
+    if (getDbEither.isLeft) {
+      return Either.left(getDbEither.left);
+    }
 
     String selectCommand = ""
         "SELECT "
@@ -92,17 +115,25 @@ class SqfliteServiceRepository implements ServiceRepository {
         "$servicesTable ser "
         "WHERE ser.id = ?";
 
-    List params = [id];
-    List<Map> serviceMap = await database!.rawQuery(selectCommand, params);
-
-    Service service = ServiceAdapter.fromSqflite(map: serviceMap[0]);
-
-    return service;
+    try {
+      final params = [id];
+      final serviceMap = await database!.rawQuery(selectCommand, params);
+      final service = ServiceAdapter.fromSqflite(map: serviceMap[0]);
+      return Either.right(service);
+    } on DatabaseException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao capturar dados locais: $e'));
+    } on FileSystemException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao capturar dados locais: ${e.message})'));
+    }
   }
 
   @override
-  Future<List<Service>> getNameContained({required String name}) async {
-    await _initDatabase();
+  Future<Either<Failure, List<Service>>> getNameContained({required String name}) async {
+    final getDbEither = await _initDatabase();
+    if (getDbEither.isLeft) {
+      return Either.left(getDbEither.left);
+    }
+
     String nameWithoutDiacritic = replaceDiacritic(name).trim().toUpperCase();
 
     String selectCommand = ""
@@ -120,23 +151,28 @@ class SqfliteServiceRepository implements ServiceRepository {
         "WHERE "
         "trim(upper(ser.nameWithoutDiacritic)) LIKE '%$nameWithoutDiacritic%'";
 
-    List<Map> serviceMap = await database!.rawQuery(selectCommand);
-
-    List<Service> services = serviceMap
-        .map((service) => ServiceAdapter.fromSqflite(map: service))
-        .toList();
-
-    return services;
+    try {
+      final serviceMap = await database!.rawQuery(selectCommand);
+      final services = serviceMap.map((service) => ServiceAdapter.fromSqflite(map: service)).toList();
+      return Either.right(services);
+    } on DatabaseException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao capturar dados locais: $e'));
+    } on FileSystemException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao capturar dados locais: ${e.message})'));
+    }
   }
 
   @override
-  Future<List<Service>> getUnsync({required DateTime dateLastSync}) {
-    throw UnimplementedError();
+  Future<Either<Failure, List<Service>>> getUnsync({required DateTime dateLastSync}) async {
+    return Either.left(Failure('Método não desenvolvido'));
   }
 
   @override
-  Future<String> insert({required Service service}) async {
-    await _initDatabase();
+  Future<Either<Failure, String>> insert({required Service service}) async {
+    final getDbEither = await _initDatabase();
+    if (getDbEither.isLeft) {
+      return Either.left(getDbEither.left);
+    }
 
     String insert = ""
         "INSERT INTO $servicesTable "
@@ -152,7 +188,7 @@ class SqfliteServiceRepository implements ServiceRepository {
         ") "
         "VALUES (?, ?, ?, ?, ?, ?, ?, ?)";
 
-    List params = [
+    final params = [
       service.id,
       service.serviceCategoryId,
       service.name.trim(),
@@ -163,14 +199,22 @@ class SqfliteServiceRepository implements ServiceRepository {
       service.nameWithoutDiacritics.trim(),
     ];
 
-    await database!.rawInsert(insert, params);
-
-    return service.id;
+    try {
+      await database!.rawInsert(insert, params);
+      return Either.right(service.id);
+    } on DatabaseException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao inserir dados locais: $e'));
+    } on FileSystemException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao inserir dados locais: ${e.message})'));
+    }
   }
 
   @override
-  Future<void> update({required Service service}) async {
-    await _initDatabase();
+  Future<Either<Failure, Unit>> update({required Service service}) async {
+    final getDbEither = await _initDatabase();
+    if (getDbEither.isLeft) {
+      return Either.left(getDbEither.left);
+    }
 
     String updateText = ''
         'UPDATE $servicesTable '
@@ -196,27 +240,46 @@ class SqfliteServiceRepository implements ServiceRepository {
       service.id,
     ];
 
-    await database!.rawUpdate(updateText, params);
+    try {
+      await database!.rawUpdate(updateText, params);
+      return Either.right(unit);
+    } on DatabaseException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao atualizar dados locais: $e'));
+    } on FileSystemException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao atualizar dados locais: ${e.message})'));
+    }
   }
 
   @override
-  Future<void> deleteById({required String id}) async {
-    await _initDatabase();
+  Future<Either<Failure, Unit>> deleteById({required String id}) async {
+    final getDbEither = await _initDatabase();
+    if (getDbEither.isLeft) {
+      return Either.left(getDbEither.left);
+    }
 
     String deleteTextMold = ''
         'DELETE FROM '
         '$servicesTable '
         'WHERE '
         'id = ?';
+    final params = [id];
 
-    List params = [id];
-
-    await database!.rawDelete(deleteTextMold, params);
+    try {
+      await database!.rawDelete(deleteTextMold, params);
+      return Either.right(unit);
+    } on DatabaseException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao apagar dados locais: $e'));
+    } on FileSystemException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao apagar dados locais: ${e.message})'));
+    }
   }
 
   @override
-  Future<bool> existsById({required String id}) async {
-    await _initDatabase();
+  Future<Either<Failure, bool>> existsById({required String id}) async {
+    final getDbEither = await _initDatabase();
+    if (getDbEither.isLeft) {
+      return Either.left(getDbEither.left);
+    }
 
     String selectCommand = ""
         "SELECT "
@@ -226,10 +289,15 @@ class SqfliteServiceRepository implements ServiceRepository {
         "WHERE "
         "ser.id = ?";
 
-    List params = [id];
-    List<Map> serviceMap =
-        await database!.rawQuery(selectCommand, params);
+    final params = [id];
 
-    return serviceMap.isNotEmpty;
+    try {
+      final serviceMap = await database!.rawQuery(selectCommand, params);
+      return Either.right(serviceMap.isNotEmpty);
+    } on DatabaseException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao capturar dados locais: $e'));
+    } on FileSystemException catch (e) {
+      return Either.left(GetDatabaseFailure('Falha ao capturar dados locais: ${e.message})'));
+    }
   }
 }
